AM updates 4.1
The first sentence was really long, so I chopped off the end of it. If you'd like to add it back in, I'd suggest making it it's own sentence. You also mentioned Cloudant twice, but hyperlinked it on the second reference. I just moved it to the first, as that would make more sense. Same thing with TensorFlow. It was linked in both references so I took out the second one. Struggled with creating uniformity with the lists under use case number 2 (data-at-rest through streaming data). Willing to discuss if you aren't a fan of it.